Revisions for generic ack/nak/retraction
Notes from Chad:

After much discussion with Lyteck and Brian, here’s the latest Loan FpML schema.  Note that we now have one acknowledgement, one exception, and one retraction (‘retracted) message, all built more generically, and relying on identifier with optional ‘blob’ for the original message (ack and except had the blobs already, so we added one to the generic retraction).

The only set of message that this doesn’t exactly allow us to retract are the two trade-related confirmation messages, which I think it problematic.  So, we may need to consider the construction of  a ‘confirmation identifier’ in order to include it in the identifiers model.